Videos command moved to youtube.py
